Created interfaces for CalendarImporter, CalendarService and ConfigurationService. Moved list of google calendar ids to DefaultCalendarService.
