+++ src/main/kotlin/org/steffeleffe/calendarimporter/CalendarImporter.kt
@@ -1,0 +1,9 @@
+package org.steffeleffe.calendarimporter
+
+import org.steffeleffe.calendarservice.CalendarEvent
+
+interface CalendarImporter {
+    fun importCalender(calendarId: String, numberOfDaysToImport: Int): List<CalendarEvent>
+}
+
+
