--- src/main/kotlin/org/steffeleffe/configurationservice/ConfigurationClasses.kt
@@ -1,6 +1,0 @@
-package org.steffeleffe.configurationservice
-
-import kotlinx.css.Color
-import org.steffeleffe.calendarservice.Participant
-
-data class ColorConfiguration(val participant: Participant, val color : Color)
