--- src/test/kotlin/org/steffeleffe/calendarimport/GoogleCalendarImporterTest.kt
@@ -1,5 +1,0 @@
-package org.steffeleffe.calendarimport
-
-import org.junit.jupiter.api.Assertions.*
-
-internal class GoogleCalendarImporterTest
